Update Vignette + argument "strategy"
Added explanation in vignette that the columns in the simulation result are identical if the p-hacking was "unsuccessful". Changed argument "ambitious = T/F" to argument "strategy = ..." with the options "firstsig", "smallest.sig", and "smallest" so that more p-value selection strategies are  represented. Moved the p-value selection strategies to a new function in the "helpers.R" file to reduce the code length of the simulation functions.